requirements.txt
flask
openpyxl
Pillow
supabase
python-dotenv
